Separate utils from UI dependent code
