+++ src/autolinks/index.ts
@@ -1,1 +1,2 @@
 export * from './autolinks';
+export * from './autolinks.utils';
